requirements.txt
flask==3.0.0
werkzeug==3.0.1
google-cloud-documentai==2.26.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-storage==2.13.0
google-genai
Pillow
# Note: The 'gcloud' CLI is a system tool and cannot be installed via requirements.txt
# Install it via: https://cloud.google.com/sdk/docs/install
